Time: 5 ms (79.29%), Space: 45.8 MB (86.49%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -1,7 +1,7 @@
 class Solution {
     public boolean isPalindrome(int x) {
         int temp = x;
-        int rev = 0;
+        long rev = 0;
         if(x<0)return false;
         while(x!=0 ){
             rev = rev*10 + x%10;
